Modernize and expand public interface
Expand the public interface to include all of the low-level pkt-line
methods so that people can use them.  Allow instantiating every class by
itself.  Pull in the minInt function as a helper method so it's not
necessary to depend on Git LFS and this library can be used by others.

Finally, set the repository up for Go modules so it can be independently
tested.