Implement reading delimiter packets
Currently our read code knows nothing about delimiter packets and
rejects them as invalid.  However, even if we allowed such packets, we
couldn't distinguish them from flush packets.  We currently return nil
as the array for flush packets and an empty but non-nil array for empty
packets (those with the header 0004), but there is no other valid value
we could use to represent delimiter packets.

To help solve this problem in the cases we care about, let's add an
additional set of functions which return the packet header length so we
can distinguish between these cases and write correct code.  We also
update the existing ReadPacketList function to distinguish between empty
packets (which we also don't currently support) and flush packets, and
add tests for all of these cases.