Handle empty packets
While the Git protocol documentation recommends that implementations
SHOULD NOT emit empty packets, they are still valid packets, and we
should handle them.  Let's make sure we handle them correctly and return
a nil body for a flush or delimiter packet and non-nil body for an empty
packet and add tests for these cases.  We change our invalid packets to
contain 0003 instead.