reader: properly implement io.Reader
The io.Reader interface specifies that once we reach an EOF indicator,
the next read from the stream should return 0, io.EOF.  However, our
reader didn't work that way, causing us to attempt to read the next
packet from the underlying reader.  In some cases, such a packet won't
exist until we've responded to the one we already got, causing us to hang.

This was fixed in git-lfs/git-lfs#3902, but never ported over to here.
Let's port that change over here, and add a Reset method to allow users
to continue with the next packet if there's more data to read from the
underlying stream.